+++ src/main/java/ua/spring/tasktracker/service/UserService.java
@@ -4,18 +4,23 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import ua.spring.tasktracker.dto.user.UserCreationDTO;
 import ua.spring.tasktracker.dto.user.UserDTO;
 import ua.spring.tasktracker.dto.user.UserPageDTO;
 import ua.spring.tasktracker.dto.user.UserUpdateDTO;
+import ua.spring.tasktracker.entity.Role;
 import ua.spring.tasktracker.entity.User;
 import ua.spring.tasktracker.repository.UserRepository;
 import ua.spring.tasktracker.utils.exceptions.EmailDuplicateException;
 import ua.spring.tasktracker.utils.exceptions.UserNotFoundException;
 import ua.spring.tasktracker.utils.mapper.UserMapper;
+import ua.spring.tasktracker.utils.security.CustomUserDetails;
+import ua.spring.tasktracker.utils.security.CustomUserDetailsService;
 
+import java.security.Principal;
 import java.util.List;
 
 @Slf4j
@@ -41,7 +46,8 @@ public class UserService {
         return userMapper.toDTO(userRepository.findById(id).orElseThrow(UserNotFoundException::new));
     }
 
-    public UserDTO updateUser(UserUpdateDTO user, Long id) {
+    public UserDTO updateUser(UserUpdateDTO user, Long id, Principal principal) {
+        validateAccess(id, principal);
         if (emailAlreadyInUse(user.getEmail())) {
             log.error("Email {} is already in use", user.getEmail());
             throw new EmailDuplicateException();
@@ -54,7 +60,8 @@ public class UserService {
         return userMapper.toDTO(userRepository.save(dbUser));
     }
 
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id, Principal principal) {
+        validateAccess(id, principal);
         userRepository.findById(id).orElseThrow(UserNotFoundException::new);
         log.info("Deleting user with id {}", id);
         userRepository.deleteById(id);
@@ -69,5 +76,13 @@ public class UserService {
 
     private boolean emailAlreadyInUse(String email) {
         return userRepository.existsByEmailIgnoreCase(email);
+    }
+
+    public static void validateAccess(Long userId, Principal principal) {
+        CustomUserDetails userDetails = CustomUserDetailsService.getDetails(principal);
+        if (!userId.equals(userDetails.getId()) && userDetails.getRole() != Role.ADMIN) {
+            log.info("User id = {} try to access data of user id = {}", userDetails.getId(), userId);
+            throw new AccessDeniedException("Access denied!");
+        }
     }
 }
